requirements.txt
streamlit
pandas
numpy
PyPDF2
google.generativeai
Ipython
python-dotenv
